==> R/03_aug.qmd <==
---
title: "03_aug_data"
format:
  html:
    embed-resources: true
editor: visual
---

## Data augmentation

The manufacturer data and vaccine data contains a lot of the same information, the main differences are size. Therefore we make the code with that data set for the manufacture and world population.

If we have time will make the same for the vaccine data.

Loading libraries:

```{r}
#| echo: false

# Loading libraries
library(tidyverse)
library(dplyr)
library(readr)
```

Loading the data to be cleaned:

```{r}
#| echo: false
# Loading data
clean_manufacturer <- read_tsv(file = "../data/02_manufacturer_cleaned.tsv")
clean_vaccines <- read_tsv(file = "../data/02_vaccines_cleaned.tsv")
clean_world_population <- read_tsv(file = "../data/02_world_population_cleaned.tsv")
```

We make a simple change to vaccines:

```{r}
# Adding variable to vaccines
aug_vaccine <- clean_vaccines |> 
    mutate( # Flaging if there is missing some information
    missing_total_vaccinations = is.na(total_vaccinations),
    missing_daily_vaccinations = is.na(daily_vaccinations)) |> 
  rename(`Country` = country) # Changing name
```

Now let make a new population estimate for 2021 by using the growth rate and the 2020 population:

```{r}
# Creating new variable 
options(scipen = 999) # Removing scientific notation

aug_world_population <- clean_world_population |> 
  mutate("2021 Population" = round(`2020 Population`*`Growth Rate`,digits = 0)) |> 
  # Rounding the digits 
  select(`Country/Territory`, # Selecting relevant columns
         `Continent`,
         `2022 Population`,
         `2021 Population`, 
         `2020 Population`) |> 
  rename(`Country` = `Country/Territory`) |> 
  pivot_longer(cols = starts_with("20"), # Columns starting with 20 pivot longer
               names_to = "Year", # New variable name "Year"
               values_to = "Population") |> # Variable for value added, named "Population"
  mutate(Year = as.numeric(str_remove(Year, " Population"))) # Remove the word "Population"

```

We make changes to the manufacturer data:

```{r}
# Making a new column for joining
aug_manufacturer <- clean_manufacturer |> # Making sure that year is a column that matches the date
  mutate(Year = as.numeric(format(date, "%Y")))
```

Now we join the data:

```{r}
# Joining data
## Vaccine and world population - This will happen if we have time

## Manufacture and world population
manufacturer_world_pop <- aug_manufacturer |> 
  rename(`Country` = location) |> # Renaming location
  left_join(aug_world_population,
            by = c("Country", "Year")) |> # Joining based on Country and Year
  # Changing order of column
  select(`Country`, `date`, `Year`, `Population`, `vaccine`, `total_vaccinations`) 
```

Further manipulation, first of with manufacturer and world population data:

```{r}
# Finding the total amount of people vaccinated pr country - 2 vaccination pr person
manufacturer_world_pop <- manufacturer_world_pop |>
  mutate(`Vaccinated people pr country` = round(total_vaccinations/2,0))

# Calculating the percentage of the population who has been vaccinated
manufacturer_world_pop <- manufacturer_world_pop |> 
  mutate(`Percentage Vaccinated` =(`Vaccinated people pr country`/`Population`)*100)

```

Writing files:

```{r}
# Writing 03_aug files
## Manufacturer and world population data
write_tsv(x = manufacturer_world_pop,
          file = str_c("../data/", "03_manufacturer_world_population_aug.tsv"))
## Vaccnies and world population data
#write_tsv(x = vaccines_cleaned,
         # file = str_c("../data/", "02_vaccines_cleaned.tsv"))
```
