==> R/06_analysis_vaccine_type.qmd <==
---
title: "06_analysis_vaccine_type"
format:
  html:
    embed-resources: true
editor: visual
---

## Plots

Here i want to make a plot similar to the one in 05_analysis but this time with the different vaccine types.

```{r}
#| echo: false
library(tidyverse)
library(scales)

# Load the dataset that we've used before. 
data <- read_tsv("../data/03_manufacturer_world_population_aug.tsv")

# Prepare the data. Filtering so we get those countries we want to include in the plot. Making sure that they're somewhat near each other, so that it's possible to see them all, without some of the graphs becoming to small. 
filtered_data <- data %>%
  filter(vaccine %in% c("Moderna", "Oxford/AstraZeneca", "Sinopharm/Beijing", "Sputnik V")) %>% #Choosing the countries i want to include in the plot. 
  mutate(date = as.Date(date, format = "%Y-%m-%d")) %>% 
  #making sure that the dates are written in the same way with the same notation
  arrange(vaccine, date)


# Reorder countries based on their maximum vaccination values (lowest first )
# Reorder countries by max values, making sure that those graphs with the lowest values are visible
#filtered_data <- filtered_data %>%
 # group_by(vaccine) %>%
  #mutate(max_vaccinations = max(total_vaccinations, na.rm = TRUE)) %>%
  #ungroup() %>%
  #mutate(Country = fct_reorder(Country, max_vaccinations, .desc = TRUE))  


# Plot the lines ensuring visibility of lower-value countries
Plot <- ggplot(filtered_data, aes(x = date, y = total_vaccinations, color = vaccine)) +
  geom_line(aes(group = vaccine), size = 1, linetype = "solid") +  # Force grouping and solid lines
  labs(
    title = "Vaccination Progression by vaccine type of the world",
    x = "Date",
    y = "Total Vaccinations (all types)",
    color = "Country"
  ) +
  theme_minimal()+
  theme(legend.position = "bottom")+
   scale_y_continuous(
    labels = label_comma()) +
   scale_x_date(
    breaks = "1 month",  # Includin breaks in the x axis per month of the period
    labels = date_format("%b %Y")  # formatting the labels 
  )+
  theme(
    axis.text.x = element_text(angle = 45, hjust = 1)  # Making the labels on the x-axis stand with a 45° angle 
  )+
  scale_y_continuous(
    labels = label_comma(),  # Remove scientific notation on the y-axis
    breaks = seq(0, max(filtered_data$total_vaccinations, na.rm = TRUE), by = 20000000)  # Add breaks every 20 million 
  )
Plot

# Saving plot
ggsave("../results/vaccine_progression.jpeg", plot = Plot)
```
